improvement: make api codes concise
